Update list of imported resources on snippet list when deleting snippet
